easywork-hr/requirements.txt
aerich==0.9.1
aiomysql==0.2.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dictdiffer==0.9.0
fastapi==0.116.1
ffmpy==0.6.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iso8601==2.1.0
jwt==1.4.0
loguru==0.7.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pypika-tortoise==0.6.1
python-dotenv==1.1.1
pytz==2025.2
sniffio==1.3.1
starlette==0.47.2
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
